Fix crash on startup
Was happening becuase I was re-using the dbSource
Maybe when mapping a network response back to the
database. The source has a filter on it which
doesn't emit when the item needs updating. This is
correct for the original purpose of the source,
but not when we want to map back to the database.

Fixed by creating a new Maybe without the filter.

Closes #11