Refactor HomeNavigator injection
Currently it is injected via Dagger at VM creation.
This does not work at all when the Activity is rotated,
since the VM still be holding onto the old Activity
instance.

This commits refactors this to be based around a ViewModel
and a LiveData for each action. The ViewModel holds those
LiveDatas and the current Activity observes them. Each
fragment then retrieves the NavigatorViewModel and invokes
the LiveData's (through the interface).

Closes #47