UIs now only receive a single UiMessage to display
Previously each view state would contain a list of the queued up
messages. This means that the UI is responsible for deciding what message
to display, which is a bit too much logic in the UI for my liking.

This PR changes this so each view state contains a single UiMessage?.
The queuing is moved to UiMessageManager which means that all of the
logic is still in the VM.

We could later add some kind of message prioritisation, skip messages,
etc later without any changes to the UI.