Sync season/episodes data for all watched shows
This PR changes the fundamental feature of what a followed show
actually. Previous to this PR, it controlled most of of the
features in the app (such as syncing episode data, etc). Now
being a followed show is similar to what of metadata, and will
in the future control things like notifications.