Fix list position restoration
Contains a workaround for  https://issuetracker.google.com/issues/177245496.

Due to the way which paging loads data, it will always result
in an initial empty state, and then the loaded items. That
conflicts with the way which rememberLazyListState() and
friends store their state, as they rely on the items being
present at the time of (Android) state restoration. To
workaround this, we use a different LazyListState while the lazy
paging items are empty, and then switcheroo to the standard state
once we know we have the items available.