Reduce the scope of spotless tasks
Otherwise running Spotless on the root project will analyze 54 Gradle build files and 474 Kotlin files, which will then be re-analyzed by their modules. This is redundant but it also results in long fingerprinting times. For example, it took 18s on CI to fingerprint task inputs for an UP-TO-DATE :spotlessKotlin task (on the root project)
https://scans.gradle.com/s/sqjswckwfddis/timeline?details=oajzimjzgvapi&expanded=WyI3Il0&toggled=WyIxNzciXQ&view=by-type