Use movableContentOf
This allows us to move content from the back stack to being
the 'top item' without loss of remembered state, which is important
when the Presenter and Ui state is stored as snapshot state.

There's still an issue where the content flashes when being moved, but
seems mostly like a performance issue.