Move database operations specific dispatchers
Reads are limited to a dispatcher with a parallelism of 4, which
is equal to the newly set maximum connection readers on the SQLite
driver.

Writes are limited to a parallelism of 1, since that is what SQLDelight
will do under the hood anyway. This seems to fix concurrency issues
which result in crashes, such as https://github.com/cashapp/sqldelight/issues/4376.